Started manual mapping of DTOs found a fair amount of inconsistencies

+++ PowerLifting.Data/Entities/Templates/TemplateRepScheme.cs
@@ -4,10 +4,10 @@
     {
         public int TemplateRepSchemeId { get; set; }
         public int TemplateExerciseId { get; set; }
-        public double? Percentage { get; set; }
+        public decimal? Percentage { get; set; }
         public int SetNo { get; set; }
         public int NoOfReps { get; set; } //minimum number of reps to be performed
-        public double WeightLifted { get; set; }
+        public decimal WeightLifted { get; set; }
         public bool IsBackOffSet { get; set; }
         public bool AMRAP { get; set; } //As many reps as possible for this set
     }
